Some code style fixes

+++ kpresenter/part/KPrViewModeSlidesSorter.cpp
@@ -59,7 +59,6 @@
 
 KPrViewModeSlidesSorter::KPrViewModeSlidesSorter(KoPAView *view, KoPACanvas *canvas)
     : KoPAViewMode( view, canvas )
-    //, m_slidesSorter( new KPrSlidesSorter(this, view->parentWidget()) )
     , m_slidesSorter( new KPrSlidesManagerView(m_toolProxy, view->parentWidget()))
     , m_documentModel(new KPrSlidesSorterDocumentModel(this, view->parentWidget()))
     , m_iconSize( QSize(200, 200) )
@@ -73,7 +72,7 @@ KPrViewModeSlidesSorter::KPrViewModeSlidesSorter(KoPAView *view, KoPACanvas *can
     m_slidesSorter->installEventFilter(this);
 
     //install selection manager for Slides Sorter View
-    m_selectionManagerSlidesSorter = new KPrSelectionManager(m_slidesSorter, m_view->kopaDocument());
+    m_selectionManagerSlidesSorter = new KPrSelectionManager(m_slidesSorter);
 }
 
 KPrViewModeSlidesSorter::~KPrViewModeSlidesSorter()
@@ -411,9 +410,9 @@ void KPrViewModeSlidesSorter::slidesSorterContextMenu(QContextMenuEvent *event)
     menu.addAction(SmallIcon("document-new"), i18n("Add a new slide"), this, SLOT(addSlide()));
     menu.addAction(i18n("Delete selected slides"), this, SLOT(deleteSlide()));
 
-    menu.addAction(i18n( "Cut" ) ,this,  SLOT(editCut()));
-    menu.addAction(i18n( "Copy" ), this,  SLOT(editCopy()));
-    menu.addAction(i18n( "Paste" ), this, SLOT(editPaste()));
+    menu.addAction(i18n("Cut") ,this,  SLOT(editCut()));
+    menu.addAction(i18n("Copy"), this,  SLOT(editCopy()));
+    menu.addAction(i18n("Paste"), this, SLOT(editPaste()));
 
     menu.exec(event->globalPos());
 }
@@ -423,17 +422,11 @@ bool KPrViewModeSlidesSorter::eventFilter(QObject *watched, QEvent *event)
     if (watched == m_slidesSorter) {
         switch (event->type()) {
             case QEvent::KeyPress: {
-                QKeyEvent *keyEv = static_cast<QKeyEvent *>(event);
-                switch (keyEv->key()) {
-                    case Qt::Key_Delete: {
-                        deleteSlide();
-                        break;
-                    }
+                QKeyEvent *keyEv = dynamic_cast<QKeyEvent *>(event);
 
-                    default:
-                       break;
+                if (keyEv && keyEv->key() == Qt::Key_Delete) {
+                    deleteSlide();
                 }
-                break;
             }
 
             default:
